update scripts & package.json
