+++ scripts/generate.js
@@ -1,159 +1,52 @@
 // @ts-check
-// This script is use to generate various files to be used at runtime:
-// * types/schema/index.d.ts - Union type for all the JsonSchemas
-// * types/proto/index.d.ts - Union type for all the ProtobufSchemas
-// * types/proto/index.js - Exports all protobufs from one file
-// * types/index.d.ts - re-exports JSONSchema and ProtobufSchema types for better importing
-
-import fs from 'node:fs'
+import fs from 'fs'
 import path from 'path'
-import { URL } from 'url'
-import glob from 'glob-promise'
-import { formatSchemaType } from '../utils.js'
+import { mkdirp } from 'mkdirp'
+import { rimraf } from 'rimraf'
+import { execSync } from 'child_process'
 
-const __dirname = new URL('.', import.meta.url).pathname
+import { parseConfig } from './lib/parse-config.js'
+import { generateProtoTypes } from './lib/proto-types.js'
+import { PROJECT_ROOT } from './lib/utils.js'
+import { generateConfig } from './lib/generate-config.js'
+import { readJSONSchema } from './lib/read-json-schema.js'
+import { generateValidations } from './lib/generate-validations.js'
+import { generateJSONSchemaTS } from './lib/generate-jsonschema-ts.js'
 
-const readJSON = (f) =>
-  JSON.parse(fs.readFileSync(new URL(f, import.meta.url)).toString())
-/**
- * @param {string} p
- * @returns {{schemaType: String, schemaVersion: Number, schema:Object}}
- */
-const loadSchema = (p) => {
-  const { dir, name } = path.parse(p)
-  return {
-    // we get the type of the schema from the directory
-    schemaType: dir.replace('/tmp/schema/', ''),
-    // we get the version from the filename
-    schemaVersion: parseInt(name.replace('v', '')),
-    schema: readJSON(p),
-  }
+const DIST_DIRNAME = path.join(PROJECT_ROOT, 'dist')
+const TYPES_DIRNAME = path.join(PROJECT_ROOT, 'types')
+rimraf.sync(DIST_DIRNAME)
+rimraf.sync(TYPES_DIRNAME)
+
+mkdirp.sync(DIST_DIRNAME)
+mkdirp.sync(path.join(TYPES_DIRNAME, 'proto'))
+mkdirp.sync(path.join(TYPES_DIRNAME, 'schema'))
+
+execSync('buf generate ./proto', { cwd: PROJECT_ROOT })
+
+const config = parseConfig()
+
+const protoTypesFile = generateProtoTypes(config)
+fs.writeFileSync(
+  path.join(PROJECT_ROOT, 'types/proto/types.ts'),
+  protoTypesFile
+)
+
+const configFile = generateConfig(config)
+fs.writeFileSync(path.join(PROJECT_ROOT, 'config.ts'), configFile)
+
+const jsonSchemas = readJSONSchema(config)
+
+const validationCode = generateValidations(config, jsonSchemas)
+fs.writeFileSync(path.join(DIST_DIRNAME, 'schemas.js'), validationCode)
+
+const jsonSchemaTSDefs = await generateJSONSchemaTS(config, jsonSchemas)
+for (const [filenameBase, ts] of Object.entries(jsonSchemaTSDefs)) {
+  const filepath = path.join(
+    PROJECT_ROOT,
+    'types',
+    'schema',
+    filenameBase + '.ts'
+  )
+  fs.writeFileSync(filepath, ts)
 }
-
-const schemas = glob
-  .sync('/tmp/schema/*/*.json', { cwd: 'scripts' })
-  .map(loadSchema)
-
-const latestSchemaVersions = schemas.reduce(
-  (acc, { schemaVersion, schemaType }) => {
-    if (!acc[schemaType]) {
-      acc[schemaType] = schemaVersion
-    } else {
-      if (acc[schemaType] < schemaVersion) {
-        acc[schemaType] = schemaVersion
-      }
-    }
-    return acc
-  },
-  {}
-)
-
-// types/schema/index.d.ts
-const jsonSchemaType = `
-${schemas
-  .map(
-    /** @param {Object} schema */
-    ({ schemaVersion, schemaType }) => {
-      const varName = `${formatSchemaType(schemaType)}_${schemaVersion}`
-      return `import { ${formatSchemaType(
-        schemaType
-      )} as ${varName} } from './${schemaType}/v${schemaVersion}'`
-    }
-  )
-  .join('\n')}
-
-export type JSONSchema = (${schemas
-  .map(
-    /** @param {Object} schema */
-    ({ schemaVersion, schemaType }) =>
-      `${formatSchemaType(schemaType)}_${schemaVersion}`
-  )
-  .join(' | ')})
-${schemas
-  .map(({ schemaType, schemaVersion }) => {
-    const as =
-      latestSchemaVersions[schemaType] !== schemaVersion
-        ? `as ${formatSchemaType(schemaType)}_${schemaVersion}`
-        : ''
-    return `export { ${formatSchemaType(
-      schemaType
-    )} ${as} } from './${schemaType}/v${latestSchemaVersions[schemaType]}'`
-  })
-  .join('\n')}`
-fs.writeFileSync(
-  path.join(__dirname, '../types/schema/index.d.ts'),
-  jsonSchemaType
-)
-
-// types/proto/index.d.ts and types/proto/index.js
-const protobufFiles = glob.sync('../types/proto/*/*.ts', { cwd: 'scripts' })
-const obj = protobufFiles
-  .filter((f) => !f.match(/.d.ts/))
-  .filter((f) => !f.match(/common/))
-  .map((p) => {
-    const { name, dir } = path.parse(p)
-    return {
-      schemaType: dir.replace('../types/proto/', ''),
-      schemaVersion: name,
-    }
-  })
-
-const linesjs = []
-const linesdts = []
-const union = obj
-  .map(
-    ({ schemaType, schemaVersion }) =>
-      `${formatSchemaType(schemaType)}_${schemaVersion.replace('v', '')}`
-  )
-  .join(' | ')
-
-const individualExports = schemas
-  .map(
-    ({ schemaType, schemaVersion }) =>
-      `export { ${formatSchemaType(schemaType)}_${schemaVersion} ${
-        latestSchemaVersions[schemaType] === schemaVersion
-          ? `as ${formatSchemaType(schemaType)}`
-          : ''
-      }} from './${schemaType}/v${schemaVersion}'`
-  )
-  .join('\n')
-
-obj.forEach(({ schemaType, schemaVersion }) => {
-  const linejs = `export { ${formatSchemaType(
-    schemaType
-  )}_${schemaVersion.replace(
-    'v',
-    ''
-  )} } from './${schemaType}/${schemaVersion}.js'`
-
-  const linedts = `import { ${formatSchemaType(
-    schemaType
-  )}_${schemaVersion.replace(
-    'v',
-    ''
-  )} } from './${schemaType}/${schemaVersion}'`
-  linesdts.push(linedts)
-  linesjs.push(linejs)
-})
-
-fs.writeFileSync(
-  path.join(__dirname, '../types/proto/index.js'),
-  linesjs.join('\n')
-)
-fs.writeFileSync(
-  path.join(__dirname, '../types/proto/index.d.ts'),
-  `${linesdts.join('\n')}
-export type ProtobufSchema = ${union}
-${individualExports}`
-)
-
-// types/index.d.ts
-fs.writeFileSync(
-  path.join(__dirname, '../types/index.d.ts'),
-  `export { ProtobufSchema } from './proto'
-export { JSONSchema } from './schema'
-export type schemaType =  ${schemas
-    .map(({ schemaType }) => `'${schemaType}'`)
-    .join(' | ')} | string
-`
-)
