--- scripts/jsonts
@@ -1,12 +1,0 @@
-#!/usr/bin/env bash
-
-for file in /tmp/schema/**/*.json
-do
-  name="${file%%.*}"
-  p=${name#/tmp/}
-  # do not generate files for common.json (since we're embeding it)
-  # if [[ $name != *common* ]];then
-  mkdir -p types/$(dirname $p)
-  cat "$file" | json2ts --cwd 'schema/*' > "types/$p.ts"
-  # fi
-done
