+++ tsconfig.json
@@ -15,6 +15,6 @@
     "declaration": true,
     "declarationMap": false
   },
-  "include": ["types/proto/*.ts", "types/schema/*.d.ts"],
+  "include": ["types/proto/*.ts", "types/schema/*.ts"],
   "exclude": ["node_modules"]
 }
